edualloc-pro/backend/requirements.txt
# ── Web Framework ────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Validation ───────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Logging ──────────────────────────────────────────────────────
structlog==24.1.0

# ── Google Cloud ─────────────────────────────────────────────────
google-cloud-bigquery==3.21.0
google-cloud-aiplatform==1.52.0
google-generativeai==0.5.4
google-auth==2.29.0
firebase-admin==6.5.0

# ── AI / Optimization ────────────────────────────────────────────
ortools==9.15.6755
numpy>=2.0.2

# ── PDF ──────────────────────────────────────────────────────────
reportlab==4.1.0

# ── HTTP ─────────────────────────────────────────────────────────
httpx==0.27.0

# ── Testing ──────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0
httpx==0.27.0

# ── Dev tools ────────────────────────────────────────────────────
python-dotenv==1.0.1
